fix: modified createTable functions
