+++ src/helpers/testHelpers.js
@@ -2,6 +2,8 @@ const os = require("os");
 const fs = require("fs");
 const path = require("path");
 const packPath = require("package-json-path");
+const Table = require("cli-table");
+
 
 /**
  * 
@@ -461,10 +463,157 @@ const createReportHTML = async (suiteIdentifier, virtualUser, testOptions, testU
   }
 };
 
+/**
+ *
+ * @param {string | number} suiteIdentifier Test Suite Identifier
+ * @param {number} virtualUser Test run number
+ * @param {string} testUuid Running test identifier
+ * @param {string} rootPath Path of the root folder of the project from which it is called
+ * @param {Array} columnNames Name of table headers
+ * @param {string} reportMode Type of report mode ( dynamicDeep | normal )
+ * 
+ * @description Print the report table
+ * @returns Table
+ */
+const createTable = async (suiteIdentifier, virtualUser, testUuid, rootPath, columnNames, reportMode) => {
+  const jestOutput = require(path.join(rootPath, `${suiteIdentifier}-jest-output.json`));
+
+  let date = new Date();
+
+  console.info(
+    `\n# ${virtualUser} Jest report : ${suiteIdentifier}\n`
+      .yellow.bold
+  );
+
+  console.info(
+    `Test uuid: ${testUuid}\n`
+      .yellow.bold
+  );
+
+  console.info(
+    `Test timestamp: ${date.toString()}\n`
+      .yellow.bold
+  );
+
+  const tableHead = [];
+  const colWidths = [];
+
+  tableHead.push("#".blue);
+  colWidths.push(5);
+
+  columnNames.forEach((column, index) => {
+    let columnWidth = 35;
+
+    if (reportMode === "dynamicDeep")
+      columnWidth = index !== 1 ? 35 : 50
+
+
+    tableHead.push(`${column}`.blue);
+    colWidths.push(columnWidth);
+  });
+
+  tableHead.push("Status".blue);
+  colWidths.push(15);
+
+  const table = new Table({
+    head: [...tableHead],
+    colWidths: [...colWidths],
+    colors: true,
+  }); //* Creates the table
+
+  let testResultIndex = 0;
+
+  let testResults = sortTestResults(jestOutput.testResults)
+
+  for (const testResult of testResults) {
+    const path =
+      os.type() === "Windows_NT"
+        ? testResult.name.split("\\")
+        : testResult.name.split("/");
+
+    const testIndex = path.indexOf("tests");
+
+    if (testIndex === -1) {
+      console.log(
+        `${path[path.length - 1]} test is not inside the correct directory.`
+          .yellow
+      );
+      testResultIndex++;
+      continue;
+    }
+
+    let tableValues = path.slice(testIndex + 1, path.length);
+
+    if (tableValues.length !== columnNames.length && reportMode !== 'dynamicDeep') {
+      console.log(
+        `${path[path.length - 1]} does not meet your columns definition.`.yellow
+      );
+    }
+
+    /**
+     * If the type of report is dynamic, adjust depth of folders in columns
+     */
+    if (reportMode === 'dynamicDeep') {
+      let fixedColumns = [];
+
+      /**
+       * The first is always used for column 'C1'
+       */
+
+      fixedColumns.push((tableValues.length > 1 && columnNames.length > 1) ? tableValues.shift() : "");
+
+      /**
+       * Iterate and concatenate the folder to fixed
+       */
+      let dynamicColumn = ''
+      for (let index = 0; index < tableValues.length - 1; index++) {
+        index === 0 ? '' : dynamicColumn += '/';
+        dynamicColumn += `${tableValues[index]}`
+      }
+
+      /**
+       * Add the fixed column 'C2'
+       */
+      fixedColumns.push(dynamicColumn)
+      
+      /**
+       * Add the value of the last column 'C3'
+       */
+      fixedColumns.push(tableValues.pop().split('.test')[0])
+
+      // Replace table value
+      tableValues = fixedColumns
+    }
+
+    const contentToPush = [];
+
+    contentToPush.push((testResultIndex + 1).toString());
+
+    for (let index = 0; index < columnNames.length; index++) {
+      if (tableValues[index]) {
+        contentToPush.push(tableValues[index]);
+      } else {
+        contentToPush.push("");
+      }
+    }
+
+    contentToPush.push(
+      testResult.status === "passed"
+        ? testResult.status.green
+        : testResult.status.red
+    );
+
+    table.push([...contentToPush]);
+    testResultIndex++;
+  } //* Inserts data to the table
+  return table;
+};
+
 module.exports = {
   formatVarsEnv,
   getVariable,
   sortTestResults,
   takeScreenshot,
-  createReportHTML
+  createReportHTML,
+  createTable
 }
